requirements.txt
supabase>=2.7.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-dateutil>=2.8.2